chore: rename package, move some info logs to debug level
