+++ store/precomputed_key/weave_vm/types/signer.go
